+++ bob-common/src/metrics/mod.rs
@@ -1,15 +1,23 @@
 use crate::configs::node::{Node as NodeConfig, LOCAL_ADDRESS, METRICS_NAME, NODE_NAME};
+use metrics::{register_counter, Recorder};
+use metrics_exporter_prometheus::PrometheusRecorder;
+use metrics_util::MetricKindMask;
 use pearl::init_pearl;
 pub use snapshot::SharedMetricsSnapshot;
 use std::{
+    net::SocketAddr,
     sync::Arc,
     time::{Duration, Instant},
 };
 
 pub mod accumulator;
-pub mod exporter;
+mod exporters;
 pub mod pearl;
 pub mod snapshot;
+
+use exporters::global_exporter::GlobalRecorder;
+
+use self::exporters::graphite_exporter::MetricsRecorder;
 
 /// Counts number of PUT requests, processed by Grinder
 pub const GRINDER_PUT_COUNTER: &str = "grinder.put_count";
@@ -95,9 +103,9 @@ impl BobClient {
 
     #[allow(clippy::cast_possible_truncation)]
     pub(crate) fn put_timer_stop(&self, timer: Timer) {
-        timing!(
+        histogram!(
             self.prefix.clone() + ".put_timer",
-            timer.elapsed().as_nanos() as u64
+            timer.elapsed().as_nanos() as f64
         );
     }
 
@@ -110,9 +118,9 @@ impl BobClient {
     }
 
     pub(crate) fn get_timer_stop(&self, timer: Timer) {
-        timing!(
+        histogram!(
             self.prefix.clone() + ".get_timer",
-            timer.elapsed().as_nanos() as u64
+            timer.elapsed().as_nanos() as f64
         );
     }
 
@@ -129,9 +137,9 @@ impl BobClient {
     }
 
     pub(crate) fn exist_timer_stop(&self, timer: Timer) {
-        timing!(
+        histogram!(
             self.prefix.clone() + ".exist_timer",
-            timer.elapsed().as_nanos() as u64
+            timer.elapsed().as_nanos() as f64
         );
     }
 }
@@ -162,7 +170,7 @@ impl ContainerBuilder for MetricsContainer {
 }
 
 /// initializes bob counters with given config and address of the local node
-/// returns metrics and readable metrics snapshot
+#[allow(unused_variables)]
 pub fn init_counters(
     node_config: &NodeConfig,
     local_address: &str,
@@ -170,17 +178,9 @@ pub fn init_counters(
     Arc<dyn ContainerBuilder + Send + Sync>,
     SharedMetricsSnapshot,
 ) {
-    let prefix_pattern = node_config
-        .metrics()
-        .prefix()
-        .map_or(format!("{}.{}", NODE_NAME, LOCAL_ADDRESS), str::to_owned);
-    let prefix = resolve_prefix_pattern(prefix_pattern, node_config, local_address);
-    let shared = exporter::GraphiteBuilder::new()
-        .set_address(node_config.metrics().graphite().to_string())
-        .set_interval(Duration::from_secs(1))
-        .set_prefix(prefix)
-        .install()
-        .expect("Can't install metrics");
+    //install_prometheus();
+    //install_graphite(node_config, local_address);
+    let shared = install_global(node_config, local_address);
     let container = MetricsContainer::new(Duration::from_secs(1), CLIENTS_METRICS_DIR.to_owned());
     info!(
         "metrics container initialized with update interval: {}ms",
@@ -194,6 +194,56 @@ pub fn init_counters(
     (metrics, shared)
 }
 
+fn init_grinder() {
+    register_counter!(GRINDER_GET_COUNTER);
+    register_counter!(GRINDER_PUT_COUNTER);
+    register_counter!(GRINDER_EXIST_COUNTER);
+    register_counter!(GRINDER_GET_ERROR_COUNT_COUNTER);
+    register_counter!(GRINDER_PUT_ERROR_COUNT_COUNTER);
+    register_counter!(GRINDER_EXIST_ERROR_COUNT_COUNTER);
+}
+
+fn init_backend() {
+    register_counter!(BACKEND_STATE);
+    register_counter!(BLOBS_COUNT);
+    register_counter!(ALIEN_BLOBS_COUNT);
+}
+
+fn init_link_manager() {
+    register_counter!(AVAILABLE_NODES_COUNT);
+}
+
+fn install_global(node_config: &NodeConfig, local_address: &str) -> SharedMetricsSnapshot {
+    let (graphite_rec, shared) = build_graphite(node_config, local_address);
+    let prometheus_rec = build_prometheus();
+    let recorders: Vec<Box<dyn Recorder>> = vec![Box::new(graphite_rec), Box::new(prometheus_rec)];
+    install_global_recorder(recorders);
+    shared
+}
+
+fn install_global_recorder(recorders: Vec<Box<dyn Recorder>>) {
+    let global_rec = GlobalRecorder::new(recorders);
+    metrics::set_boxed_recorder(Box::new(global_rec)).expect("Can't set global recorder");
+}
+
+#[allow(unused)]
+fn install_prometheus() {
+    let recorder = build_prometheus();
+    metrics::set_boxed_recorder(Box::new(recorder)).expect("Can't set Prometheus recorder");
+}
+
+fn build_prometheus() -> PrometheusRecorder {
+    metrics_exporter_prometheus::PrometheusBuilder::new()
+        .listen_address(
+            "0.0.0.0:9000"
+                .parse::<SocketAddr>()
+                .expect("Bad metrics address"),
+        )
+        .idle_timeout(MetricKindMask::ALL, Some(Duration::from_secs(2)))
+        .build()
+}
+
+#[allow(unused)]
 fn resolve_prefix_pattern(
     mut pattern: String,
     node_config: &NodeConfig,
@@ -213,21 +263,24 @@ fn resolve_prefix_pattern(
     pattern
 }
 
-fn init_grinder() {
-    counter!(GRINDER_GET_COUNTER, 0);
-    counter!(GRINDER_PUT_COUNTER, 0);
-    counter!(GRINDER_EXIST_COUNTER, 0);
-    counter!(GRINDER_GET_ERROR_COUNT_COUNTER, 0);
-    counter!(GRINDER_PUT_ERROR_COUNT_COUNTER, 0);
-    counter!(GRINDER_EXIST_ERROR_COUNT_COUNTER, 0);
+#[allow(unused)]
+fn install_graphite(node_config: &NodeConfig, local_address: &str) {
+    let (recorder, shared) = build_graphite(node_config, local_address);
+    metrics::set_boxed_recorder(Box::new(recorder)).expect("Can't set graphite recorder");
 }
 
-fn init_backend() {
-    counter!(BACKEND_STATE, 0);
-    counter!(BLOBS_COUNT, 0);
-    counter!(ALIEN_BLOBS_COUNT, 0);
-}
-
-fn init_link_manager() {
-    counter!(AVAILABLE_NODES_COUNT, 0);
+fn build_graphite(
+    node_config: &NodeConfig,
+    local_address: &str,
+) -> (MetricsRecorder, SharedMetricsSnapshot) {
+    let prefix_pattern = node_config
+        .metrics()
+        .prefix()
+        .map_or(format!("{}.{}", NODE_NAME, LOCAL_ADDRESS), str::to_owned);
+    let prefix = resolve_prefix_pattern(prefix_pattern, node_config, local_address);
+    exporters::graphite_exporter::GraphiteBuilder::new()
+        .set_address(node_config.metrics().graphite().to_string())
+        .set_interval(Duration::from_secs(1))
+        .set_prefix(prefix)
+        .build()
 }